backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
python-dotenv
pydantic
pydantic-settings
psycopg2-binary; sys_platform != 'win32'
